==> notebooks/01_get_dataframes.ipynb <==
# Import libraries
import pandas as pd

#Url data
url = '../data/raw/PRECIOS LUZ.xlsx'

# Extracting dataframe and preprocessing
df_fixed_price = pd.read_excel(url, sheet_name='FIJO')
df_fixed_price = df_fixed_price.iloc[1:,2:]
df_fixed_price.dropna(inplace=True)
df_fixed_price.columns = ['SISTEMA','CIA','PRODUCTO','PRODUCTO_CIA','TARIFA', 'FEE',
                          'P1','P2','P3','P4','P5','P6',
                          'P1.','P2.','P3.','P4.','P5.','P6.'
                            ]

# Nice
df_fixed_price

# Saving to csv
df_fixed_price.to_csv('../data/processed/fixed_price.csv', index=False)

# Extracting dataframe df_indexed_price and preprocessing
df_indexed_price = pd.read_excel(url, sheet_name='INDEXADO')

# Proper data and nice columns
df_indexed_price = df_indexed_price.iloc[3:,1:12]
df_indexed_price.columns=['SISTEMA','TARIFA','CIA','MES','FEE','P1','P2','P3','P4','P5','P6']
df_indexed_price.reset_index(drop=True, inplace=True)

df_indexed_price

# Saving to csv
df_indexed_price.dropna(inplace=True)
df_indexed_price.reset_index(drop=True, inplace=True)
df_indexed_price.to_csv('./data/processed/indexed_price.csv', index=False)

# Extracting dataframe df_fixed_price_power and preprocessing

df_indexed_price_power = pd.read_excel(url, sheet_name='INDEXADO')

# Making everything nice
df_indexed_price_power = df_indexed_price_power.iloc[3:, 13:]
df_indexed_price_power.columns = ['SISTEMA','CIA','PRODUCTO','PRODUCTO_CIA','TARIFA','P1','P2','P3','P4','P5','P6']
df_indexed_price_power.dropna(inplace=True)
df_indexed_price_power.reset_index(drop=True, inplace=True)

df_indexed_price_power

# Saving to csv
df_indexed_price_power.to_csv('../data/processed/indexed_price_power.csv', index=False)

url1 = './data/raw/ESTUDIOPERSONALIZADO_SEVERAL_LUZ (2).xlsx'

df_indexed1 = pd.read_excel(url1, sheet_name='INDEXADO')

df_indexed1 = df_indexed1.iloc[2:,2:13]
df_indexed1.dropna(inplace=True)
df_indexed1.reset_index(drop=True, inplace=True)
df_indexed1.columns = df_indexed1.iloc[0]
df_indexed1 = df_indexed1[1:]


df_indexed1

# Saving to csv
df_indexed1.to_csv('../data/processed/indexed1.csv', index=False)
