==> notebooks/03_inicio_formulas.ipynb <==
import numpy as np
import pandas as pd
import psycopg2

db_params = {
    'host': '34.78.249.103',
    'database': 'postgres',
    'user': 'postgres',
    'password': 'cristian99'
}

conn = psycopg2.connect(**db_params)
cursor = conn.cursor()


consulta_fixed = "SELECT * FROM fixed_price"
consulta_indexed = "SELECT * FROM indexed_price"
consulta_power = "SELECT * FROM indexed_price_power"

df_fixed = pd.read_sql_query(consulta_fixed, conn)
index_price = pd.read_sql_query(consulta_indexed, conn)
index_power = pd.read_sql_query(consulta_power, conn)
index_price_anual = index_price.copy()
index_price_power_anual = index_power.copy()

cursor.close()
conn.close()

display(index_price)

## Factura de prueba (kino)

cons_mens_P1= 74 #punta #kW
cons_mens_P2= 83 #llano 
cons_mens_P3= 168 #valle

sumatorio_cons_mens=cons_mens_P1+cons_mens_P2+cons_mens_P3

precio_mens_P1= 0.128515 # €
precio_mens_P2= 0.128515
precio_mens_P3= 0.128515

descuento= 0.02 #porcentaje

precio_P1_descuento= precio_mens_P1 * (1-descuento) #€
precio_P2_descuento= precio_mens_P2 * (1-descuento)
precio_P3_descuento= precio_mens_P3 * (1-descuento)


total_pago_P1_energia= cons_mens_P1 * precio_P1_descuento #€
total_pago_P2_energia= cons_mens_P2 * precio_P2_descuento
total_pago_P3_energia= cons_mens_P3 * precio_P3_descuento

sumatorio_total_pago_energia = total_pago_P1_energia + total_pago_P2_energia + total_pago_P3_energia

def calcular_energia(cons_mens_P1,cons_mens_P2,cons_mens_P3, precio_mens_P1,precio_mens_P2,precio_mens_P3,descuento):
    sumatorio_cons_mens=cons_mens_P1+cons_mens_P2+cons_mens_P3
    precio_P1_descuento= precio_mens_P1 * (1-descuento) #€
    precio_P2_descuento= precio_mens_P2 * (1-descuento)
    precio_P3_descuento= precio_mens_P3 * (1-descuento)

    total_pago_P1_energia= cons_mens_P1 * precio_P1_descuento #€
    total_pago_P2_energia= cons_mens_P2 * precio_P2_descuento
    total_pago_P3_energia= cons_mens_P3 * precio_P3_descuento

    sumatorio_total_pago_energia = total_pago_P1_energia + total_pago_P2_energia + total_pago_P3_energia
    return sumatorio_total_pago_energia

potencia_contratada_P1= 3.45 #kW
potencia_contratada_P2= 3.45

dias=33
precio_potencia_dia_P1= 0.09968 #€
precio_potencia_dia_P2= 0.033816
total_pago_P1_potencia= dias * precio_potencia_dia_P1 * potencia_contratada_P1
total_pago_P2_potencia= dias * precio_potencia_dia_P2 * potencia_contratada_P2

sumatorio_total_pago_potencia = total_pago_P1_potencia + total_pago_P2_potencia


def calcular_potencia(potencia_contratada_P1,potencia_contratada_P2,dias,precio_potencia_dia_P1,precio_potencia_dia_P2):
    total_pago_P1_potencia= dias * precio_potencia_dia_P1 * potencia_contratada_P1
    total_pago_P2_potencia= dias * precio_potencia_dia_P2 * potencia_contratada_P2
    sumatorio_total_pago_potencia = total_pago_P1_potencia + total_pago_P2_potencia
    return sumatorio_total_pago_potencia

impuesto_electrico=0.33 #euros

otros= 2.15 #bonos, alquiler contador
IVA = 0.05 #%
bi_IVA= (sumatorio_total_pago_energia
         +sumatorio_total_pago_potencia
         +impuesto_electrico
         +otros)
importe_total_factura_mens= bi_IVA * (1+IVA)
importe_total_factura_mens

def calcular_total_factura(sumatorio_total_pago_energia,sumatorio_total_pago_potencia,impuesto_electrico,otros,IVA):
    bi_IVA= (sumatorio_total_pago_energia + sumatorio_total_pago_potencia
         +impuesto_electrico + otros)
    importe_total_factura_mens= bi_IVA * (1+IVA)
    return importe_total_factura_mens

## Propuesta

energia_indexado=pd.read_csv('../data/processed/indexed_price.csv')
potencia_indexado=pd.read_csv('../data/processed/indexed_price_power.csv')
energia_indexado_fecha=pd.read_csv('../data/processed/indexed_updated_DM.csv')

#variables se obtienen del formulario de full stack 
cons_mens_P1= 74 #punta #kW
cons_mens_P2= 83 #llano 
cons_mens_P3= 168 #valle

sumatorio_cons_mens=cons_mens_P1+cons_mens_P2+cons_mens_P3

#ML
# precio_mens_P1=  # €
# precio_mens_P2=
# precio_mens_P3=
#obtener de cada fee de cada compañía y comparar

#descuento=  #porcentaje

precio_P1_descuento= precio_mens_P1 * (1-descuento) #€
precio_P2_descuento= precio_mens_P2 * (1-descuento)
precio_P3_descuento= precio_mens_P3 * (1-descuento)


> Recorrerse con un bucle for cada fee de cada compañía, se obtienen todos los precios con las fórmulas de arriba y se calculan los más baratos

energia_indexado_fecha['MES'] = pd.to_datetime(energia_indexado_fecha['MES'])
energia_indexado_fecha.set_index('MES', inplace=True)
df_indexed_fecha = energia_indexado_fecha.sort_index().reset_index()
# df_indexed_fecha.tail()

df_indexed_fecha

energia_indexado_fecha.info()

import psycopg2
from psycopg2 import sql

# Establece la conexión a tu base de datos
conn = psycopg2.connect(
    host="tu_host",
    database="tu_base_de_datos",
    user="tu_usuario",
    password="tu_contraseña"
)

# Crea un cursor
cursor = conn.cursor()

# Tabla y columnas a modificar
tabla = "mi_tabla"
columnas = ["p1_e", "p2_e", "p3_e", "p4_e", "p5_e", "p6_e"]

# Realiza la modificación para cada columna
for columna in columnas:
    query = sql.SQL("""
        ALTER TABLE {} 
        ALTER COLUMN {} TYPE float 
        USING {}::float;
    """).format(
        sql.Identifier(tabla),
        sql.Identifier(columna),
        sql.Identifier(columna)
    )

    cursor.execute(query)

# Confirma los cambios y cierra la conexión
conn.commit()
cursor.close()
conn.close()
