==> notebooks/04_DB_queries_cloud.ipynb <==
import psycopg2
import pandas as pd
import pymysql



#------------------------------------------------------------Conectamos y sacamos los datos de la database---------------------------------------------------------

db_params = {
    'host': '34.78.249.103',
    'database': 'postgres',
    'user': 'postgres',
    'password': 'cristian99'
}

conn = psycopg2.connect(**db_params)
cursor = conn.cursor()


consulta = "SELECT * FROM fixed_price"
cursor.execute(consulta)

resultados = cursor.fetchall()

df = pd.read_sql_query(consulta, conn)

cursor.close()
conn.close()

display(df)

import pymysql
import pandas as pd

db_params = {
    'host': '34.78.249.103',
    'database': 'postgres',
    'user': 'postgres',
    'password': 'cristian99'
}

host ='34.78.249.103'
database = 'postgres'
user = 'postgres'
password = 'cristian99'

conn = pymysql.connect(host=host, user=user, password=password, database=database,port=5432)
cursor = conn.cursor()


consulta = "SELECT * FROM fixed_price"
cursor.execute(consulta)
cursor.execute("SELECT * FROM fixed_price")
df_fixed = cursor.fetchall()
df_fixed = pd.DataFrame(df_fixed, columns=[desc[0] for desc in cursor.description])
cursor.execute("SELECT * FROM indexed_price")
consulta_indexed = cursor.fetchall()
consulta_indexed = pd.DataFrame(consulta_indexed, columns=[desc[0] for desc in cursor.description])
cursor.execute("SELECT * FROM indexed_price_power")
consulta_indexed_power = cursor.fetchall()
consulta_indexed_power = pd.DataFrame(consulta_indexed_power, columns=[desc[0] for desc in cursor.description])
# No olvides cerrar el cursor y la conexión cuando hayas terminado
cursor.close()
conn.close()



from sqlalchemy import create_engine
import pandas as pd

host ='34.78.249.103'
database = 'postgres'
user = 'postgres'
password = 'cristian99'
port = 5432

connection_string = f"mysql+pymysql://{user}:{password}@{host}:{port}/{database}"

# Crea el motor de SQLAlchemy
engine = create_engine(connection_string)

# Ejecuta las consultas y carga los resultados en DataFrames
df_fixed = pd.read_sql("SELECT * FROM fixed_price", engine)
df_indexed = pd.read_sql("SELECT * FROM indexed_price", engine)
df_indexed_power = pd.read_sql("SELECT * FROM indexed_price_power", engine)

df_fixed

df = pd.read_csv("../data/processed/fixed_price.csv")

df['P1_E'] = df['P1_E'].replace('-', '0')
df['P2_E'] = df['P2_E'].replace('-', '0')
df['P3_E'] = df['P3_E'].replace('-', '0')
df['P4_E'] = df['P4_E'].replace('-', '0')
df['P5_E'] = df['P5_E'].replace('-', '0')
df['P6_E'] = df['P6_E'].replace('-', '0')

df[['P1_E','P2_E','P3_E','P4_E','P5_E','P6_E']]=df[['P1_E','P2_E','P3_E','P4_E','P5_E','P6_E']].astype(float)

df.dropna(inplace=True)

"""df.to_csv('../data/processed/fixed_price.csv', index=False)"""

df.dtypes

df.isna().sum()

df2 = pd.read_csv("../data/processed/indexed_price.csv")

df2.isna().sum()

df2.dropna(inplace=True)

"""df2.to_csv('../data/processed/indexed_price.csv', index=False)"""

df3 = pd.read_csv("../data/processed/indexed_price_power.csv")

df3.isna().sum()
