==> notebooks/05_ML_RRNN_dffiltrados.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from sklearn.preprocessing import LabelEncoder
from tensorflow.keras.layers import Dropout

df = pd.read_csv("../data/processed/fixed_price.csv")

df

object_columns = df.select_dtypes(include='object')

label_encoder = LabelEncoder()


for column in object_columns:
    df[column] = label_encoder.fit_transform(df[column])

df.drop(["P2","P3", "P4","P5","P6","P1.","P2.","P3.","P4.","P5.","P6."], axis=1, inplace=True)


df

X = df.drop('P1', axis=1)
y = df['P1']


scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)


X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)


model = Sequential()
model.add(Dense(64, input_dim=X_train.shape[1], activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(32, activation='relu'))
model.add(Dense(1, activation='linear'))



model.compile(optimizer='adam', loss='mean_squared_error')


model.fit(X_train, y_train, epochs=50, batch_size=32, validation_data=(X_test, y_test))


loss = model.evaluate(X_test, y_test)
print(f'Error cuadrático medio en el conjunto de prueba: {loss}')


predictions = model.predict(X_test)


results = model.evaluate(X_test, y_test)
results

df2 = pd.read_csv("../data/processed/indexed_price.csv")

df2=df2.dropna()

print(df2.isna().sum())

object_columns = df2.select_dtypes(include='object')

label_encoder = LabelEncoder()


for column in object_columns:
    df2[column] = label_encoder.fit_transform(df2[column])

df2.drop(["P2","P3","P4","P5","P6","MES","FEE"], axis=1, inplace=True)
df2 = df2[df2["P1"] != 0]

X = df2.drop('P1', axis=1)
y = df2['P1']


scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)


X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)


model2 = Sequential()
model2.add(Dense(64, input_dim=X_train.shape[1], activation='relu'))
model2.add(Dropout(0.5))
model2.add(Dense(32, activation='relu'))
model2.add(Dense(1, activation='linear'))

model2.compile(optimizer='adam', loss='mean_squared_error')

model2.fit(X_train, y_train, epochs=50, batch_size=32, validation_data=(X_test, y_test))

loss = model2.evaluate(X_test, y_test)
print(f'Error cuadrático medio en el conjunto de prueba: {loss}')

predictions2 = model2.predict(X_test)

results2 = model2.evaluate(X_test, y_test)
results2

df3 = pd.read_csv("../data/processed/indexed_price_power.csv")

df3

df3=df3.dropna()

print(df3.isna().sum())

object_columns = df3.select_dtypes(include='object')

label_encoder = LabelEncoder()


for column in object_columns:
    df3[column] = label_encoder.fit_transform(df3[column])

df3.drop(["P2","P3","P4","P5","P6"], axis=1, inplace=True)
df3 = df3[df3["P1"] != 0]

X = df3.drop('P1', axis=1)
y = df3['P1']


scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)


X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)


model3 = Sequential()
model3.add(Dense(64, input_dim=X_train.shape[1], activation='relu'))
model3.add(Dropout(0.5))
model3.add(Dense(32, activation='relu'))
model3.add(Dense(1, activation='linear'))

model3.compile(optimizer='adam', loss='mean_squared_error')

model3.fit(X_train, y_train, epochs=50, batch_size=32, validation_data=(X_test, y_test))

loss = model3.evaluate(X_test, y_test)
print(f'Error cuadrático medio en el conjunto de prueba: {loss}')

predictions3 = model3.predict(X_test)

results3 = model3.evaluate(X_test, y_test)
results3

df4 = pd.read_csv("../data/processed/indexed_updated_DM.csv")

df4

df4=df4.dropna()
print(df4.isna().sum())

object_columns = df4.select_dtypes(include='object')

label_encoder = LabelEncoder()


for column in object_columns:
    df4[column] = label_encoder.fit_transform(df4[column])

df4.drop(["P2.","P3.","P4.","P5","P6."], axis=1, inplace=True)
df4 = df4[df4["P1."] != 0]

X = df4.drop('P1.', axis=1)
y = df4['P1.']


scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)


X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)


model4 = Sequential()
model4.add(Dense(64, input_dim=X_train.shape[1], activation='relu'))
model4.add(Dropout(0.5))
model4.add(Dense(32, activation='relu'))
model4.add(Dense(1, activation='linear'))

model4.compile(optimizer='adam', loss='mean_squared_error')

model4.fit(X_train, y_train, epochs=50, batch_size=32, validation_data=(X_test, y_test))

loss = model4.evaluate(X_test, y_test)
print(f'Error cuadrático medio en el conjunto de prueba: {loss}')

predictions4 = model4.predict(X_test)

results4 = model4.evaluate(X_test, y_test)
results4

df = pd.read_csv("../data/processed/fixed_price.csv")
df2 = pd.read_csv("../data/processed/indexed_price.csv")
df3 = pd.read_csv("../data/processed/indexed_price_power.csv")
df4 = pd.read_csv("../data/processed/indexed_updated_DM.csv")

df

df["TARIFA"].value_counts()

df = df[df["TARIFA"] == "2.0TD"] 

df

df.to_csv('../data/processed/2.0/fixed_price20.csv', index=False)

df2["TARIFA"].value_counts()

df2 = df2[df2["TARIFA"] == "2.0TD"]
df2.to_csv('../data/processed/2.0/indexed_price20.csv', index=False)

df3["TARIFA"].value_counts()
df3 = df3[df3["TARIFA"] == "2.0TD"]
df3.to_csv('../data/processed/2.0/indexed_price_power20.csv', index=False)

df4["TARIFA"].value_counts()
df4 = df4[df4["TARIFA"] == "2.0TD"]
df4.to_csv('../data/processed/2.0/indexed_updated_DM20.csv', index=False)
