==> notebooks/06_calculadora_anual_fijo_porcentaje.ipynb <==
import pandas as pd
df = pd.read_csv('../data/processed/fixed_price.csv')

df['FEE'].value_counts()

# df.rename(columns={'P1.':'P1_E','P2.':'P2_E','P3.':'P3_E','P4.':'P4_E','P5.':'P5_E','P6.':'P6_E'},inplace=True)
# df.rename(columns={'P1':'P1_P','P2':'P2_P','P3':'P3_P','P4':'P4_P','P5':'P5_P','P6':'P6_P'},inplace=True)

df.head()

# df.to_csv('../data/processed/fixed_price.csv',index=False)

# condiciones para filtrar (sería con el desplegable)
condiciones_sistema_tarifa = (df['SISTEMA'] == 'PENINSULA') & (df['TARIFA'] == '2.0TD')# & (df['PRODUCTO'] == 'FIJO')

# condiciones_cias = df['CIA'].isin(['ACCIONA', 'AEQ', 'CANDELA', 'FACTOR', 'IGNIS', 'MAX'])

df_filtrado = df[condiciones_sistema_tarifa] # aplicar el filtro

df_filtrado.head()

df_filtrado['CIA'].unique()

df_filtrado[['P1_E','P2_E','P3_E','P4_E','P5_E','P6_E']]=df_filtrado[['P1_E','P2_E','P3_E','P4_E','P5_E','P6_E']].astype(float)

df_filtrado.info()

# funciones de cálculo importes
def calcular_energia_anual(cons_anual_P1,cons_anual_P2,cons_anual_P3, precio_mens_P1,precio_mens_P2,precio_mens_P3,descuento):
    # sumatorio_cons_mens=cons_mens_P1+cons_mens_P2+cons_mens_P3
    # cons_anual_P1,cons_anual_P2,cons_anual_P3 ==========>TIENEN QUE VENIR DEL SCRAPPING DE CANDELA POR EL CUPS
    precio_P1_descuento= precio_mens_P1 * (1-descuento) #€
    precio_P2_descuento= precio_mens_P2 * (1-descuento)
    precio_P3_descuento= precio_mens_P3 * (1-descuento)

    total_pago_P1_energia= cons_anual_P1 * precio_P1_descuento #€
    total_pago_P2_energia= cons_anual_P2 * precio_P2_descuento
    total_pago_P3_energia= cons_anual_P3 * precio_P3_descuento

    sumatorio_total_pago_energia = total_pago_P1_energia + total_pago_P2_energia + total_pago_P3_energia
    return sumatorio_total_pago_energia

def calcular_potencia_anual(potencia_contratada_P1,potencia_contratada_P2,precio_potencia_dia_P1,precio_potencia_dia_P2,dto_p):
    
    # OJO HAY Q SCRAPEAR LAS POTENCIA_CONTRATADA
    total_pago_P1_potencia= 365 * precio_potencia_dia_P1 * potencia_contratada_P1 * (1-dto_p)
    total_pago_P2_potencia= 365 * precio_potencia_dia_P2 * potencia_contratada_P2 * (1-dto_p)
    sumatorio_total_pago_potencia = total_pago_P1_potencia + total_pago_P2_potencia
    return sumatorio_total_pago_potencia

def calcular_total_factura_anual(sumatorio_total_pago_energia,sumatorio_total_pago_potencia,impuesto_electrico,otros,alquiler_equipo,IVA):
    bi_IVA= (sumatorio_total_pago_energia + sumatorio_total_pago_potencia
         +impuesto_electrico*12 + + alquiler_equipo*12 + otros*12)
    importe_total_factura_anual= bi_IVA * (1+IVA)
    return importe_total_factura_anual

def encontrar_opcion_mas_barata_anual(df): #df_filtrado
    opciones = []

    cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834,955,1824 #scrapping
    potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #scrapping
    precio_mens_P1, precio_mens_P2, precio_mens_P3 = 0.128515, 0.128515, 0.128515
    precio_potencia_dia_P1, precio_potencia_dia_P2 = 0.09968, 0.033816
    descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
    dias = 365 # CAMBIAR A INPUT DE LA FACTURA
    dto_p = 0.0
    
    sumatorio_total_pago_energia = calcular_energia_anual(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_mens_P1, precio_mens_P2, precio_mens_P3, descuento)
    sumatorio_total_pago_potencia = calcular_potencia_anual(potencia_contratada_P1, potencia_contratada_P2, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
    impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

    # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

    importe_total_factura_anual_actual = round(calcular_total_factura_anual(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)


    for index,row in df.iterrows():
        cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834,955,1824 #scrapping
        potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #scrapping
        precio_mens_P1, precio_mens_P2, precio_mens_P3 = row['P1_E'], row['P2_E'], row['P3_E']
        precio_potencia_dia_P1, precio_potencia_dia_P2 = row['P1_P'], row['P2_P']
        descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
        dias = 365 # CAMBIAR A INPUT DE LA FACTURA
        dto_p = 0.0

        sumatorio_total_pago_energia = calcular_energia_anual(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_mens_P1, precio_mens_P2, precio_mens_P3, descuento)
        sumatorio_total_pago_potencia = calcular_potencia_anual(potencia_contratada_P1, potencia_contratada_P2, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
        impuesto_electrico, otros,alquiler_equipo, IVA = 0.33, 1.27,0.88, 0.05  # CAMBIAR A INPUT DE LA FACTURA

        importe_total_factura_anual = round(calcular_total_factura_anual(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros, alquiler_equipo,IVA),2)

        opciones.append({
            'CIA': row['CIA'],
            'FEE': row['FEE'],
            'PRODUCTO_CIA': row['PRODUCTO_CIA'],
            'CostoTotal': importe_total_factura_anual
        })

    # Opción más barata para cada compañía
    df_opciones = pd.DataFrame(opciones)
    idx_opcion_mas_barata = df_opciones['CostoTotal'].idxmin()
    
    opcion_barata=df_opciones.iloc[idx_opcion_mas_barata]

    opciones_mas_baratas = df_opciones.nsmallest(5, 'CostoTotal')
    ahorro_euros=round(importe_total_factura_anual_actual-opcion_barata['CostoTotal'],2)
    porcentaje_ahorro= round((ahorro_euros/importe_total_factura_anual_actual)*100,2)

    return ('Precio actual:',importe_total_factura_anual_actual,'Opción más barata:',opcion_barata,
            'Ahorro:',ahorro_euros, 'Porcentaje de ahorro:',f"{porcentaje_ahorro:.1f}%","Opciones más baratas:", opciones_mas_baratas)


encontrar_opcion_mas_barata_anual(df_filtrado)

def encontrar_opcion_mas_barata_cia_anual(df): #df_filtrado
    opciones = []

    cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834,955,1824 #scrapping
    potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #scrapping
    precio_mens_P1, precio_mens_P2, precio_mens_P3 = 0.128515, 0.128515, 0.128515
    precio_potencia_dia_P1, precio_potencia_dia_P2 = 0.09968, 0.033816
    descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
    dias = 365 # CAMBIAR A INPUT DE LA FACTURA
    dto_p = 0.0
    
    sumatorio_total_pago_energia = calcular_energia_anual(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_mens_P1, precio_mens_P2, precio_mens_P3, descuento)
    sumatorio_total_pago_potencia = calcular_potencia_anual(potencia_contratada_P1, potencia_contratada_P2, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
    impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

    # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

    importe_total_factura_anual_actual = round(calcular_total_factura_anual(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)


    min_cost_dict = {}

    for index,row in df.iterrows():
        cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834,955,1824 #------------->scrapping
        potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #----------------->scrapping
        precio_mens_P1, precio_mens_P2, precio_mens_P3 = row['P1_E'], row['P2_E'], row['P3_E']
        precio_potencia_dia_P1, precio_potencia_dia_P2 = row['P1_P'], row['P2_P']
        descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
        dias = 365 # CAMBIAR A INPUT DE LA FACTURA
        dto_p = 0.0
        
        sumatorio_total_pago_energia = calcular_energia_anual(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_mens_P1, precio_mens_P2, precio_mens_P3, descuento)
        sumatorio_total_pago_potencia = calcular_potencia_anual(potencia_contratada_P1, potencia_contratada_P2, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
        impuesto_electrico, otros,alquiler_equipo, IVA = 0.33, 1.27,0.88, 0.05  # CAMBIAR A INPUT DE LA FACTURA

        importe_total_factura_anual = round(calcular_total_factura_anual(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros, alquiler_equipo,IVA),2)
        print(min_cost_dict)
        # Update the minimum cost for each 'CIA'
        if row['CIA'] not in min_cost_dict or importe_total_factura_anual < min_cost_dict[row['CIA']]:
            min_cost_dict[row['CIA']] = importe_total_factura_anual

 
    opciones = [{
        'CIA': cia,
        'FEE': df.loc[df['CIA'] == cia, 'FEE'].values[0],
        'PRODUCTO_CIA': df.loc[df['CIA'] == cia, 'PRODUCTO_CIA'].values[0],
        'CostoTotal': min_cost,
        'Ahorro': round(importe_total_factura_anual_actual - min_cost, 2),
        'PorcentajeAhorro': round(((importe_total_factura_anual_actual - min_cost) / importe_total_factura_anual_actual) * 100, 2)
    } for cia, min_cost in min_cost_dict.items()]

    # Convert the list to a DataFrame
    df_opciones = pd.DataFrame(opciones)

    # Opción más barata para cada compañía
    opciones_mas_baratas = df_opciones.nsmallest(5, 'CostoTotal')

    return opciones_mas_baratas

        


resultado= encontrar_opcion_mas_barata_cia_anual(df_filtrado)

display(resultado)
