==> notebooks/06_calculadora_anual_index_porcentaje.ipynb <==
import pandas as pd
df = pd.read_csv('../data/processed/indexed_price.csv')

df.head()

from datetime import datetime

df['MES'] = pd.to_datetime(df['MES'])

# df.drop(columns=['Unnamed: 0.1',	'Unnamed: 0'],inplace=True)
# df.to_csv('../data/processed/indexed_price.csv',index=False)

pd.set_option('display.max_rows', 120)



# Calcula la fecha máxima para cada combinación única de Sistema, Tarifa, Compañía y Fee
fecha_max_por_grupo = df.groupby(['SISTEMA', 'TARIFA', 'CIA', 'FEE'])['MES'].max()

# Filtra los datos para los últimos 12 meses para cada grupo
df_ult_12_meses = pd.DataFrame()
for index, fecha_max in fecha_max_por_grupo.items():
    filtro_grupo = (
        (df['SISTEMA'] == index[0]) &
        (df['TARIFA'] == index[1]) &
        (df['CIA'] == index[2]) &
        (df['FEE'] == index[3]) &
        (df['MES'] > fecha_max - pd.DateOffset(months=12))
    )
    df_ult_12_meses = pd.concat([df_ult_12_meses, df[filtro_grupo]])

# Calcula las medias para cada conjunto único de Sistema, Tarifa, Compañía y Fee
df_medias_index_12 = df_ult_12_meses.groupby(['SISTEMA', 'TARIFA', 'CIA', 'FEE']).agg({
    'P1': 'mean',
    'P2': 'mean',
    'P3': 'mean',
    'P4': 'mean',
    'P5': 'mean',
    'P6': 'mean'
}).reset_index()

# Puedes renombrar las columnas si es necesario
#df_medias_index_12.columns = ['SISTEMA', 'TARIFA', 'CIA', 'FEE', 'P1M_E', 'P2M_E', 'P3M_E', 'P4M_E', 'P5M_E', 'P6M_E']

# Configura el número máximo de filas a mostrar
pd.set_option('display.max_rows', 120)

# Muestra los resultados
df_medias_index_12




condicion = (df_medias_index_12['SISTEMA'] == 'PENINSULA') & (df_medias_index_12['TARIFA'] == '2.0TD') # filtrar para obtener las filas más cercanas por CIA y con SISTEMA y TARIFA específicos
df_medindx12_penins_2 = df_medias_index_12[condicion]#.groupby(['SISTEMA', 'TARIFA', 'CIA'])#.apply(lambda x: x[x['diferencia'] == x['diferencia'].min()])


df_medindx12_penins_2


df_medindx12_penins_2.dropna(axis=0,inplace=True)

df_medindx12_penins_2

index_power = pd.read_csv('../data/processed/indexed_price_power.csv')

# condiciones para filtrar (sería con el desplegable)
condiciones_sistema_tarifa = (index_power['SISTEMA'] == 'PENINSULA') & (index_power['TARIFA'] == '2.0TD') & (index_power['PRODUCTO'] == 'INDEXADO')

condiciones_cias = index_power['CIA'].isin(['ACCIONA', 'AEQ', 'CANDELA', 'FACTOR', 'IGNIS', 'MAX'])

index_power_filtrado = index_power[condiciones_sistema_tarifa & condiciones_cias] # aplicar el filtro

index_power_filtrado.head()

df_medindx12_penins_2['FEE'] = df_medindx12_penins_2['FEE'].str.upper()

# funciones de cálculo importes
def calcular_energia(cons_anual_P1,cons_anual_P2,cons_anual_P3, P1M_E,P2M_E,P3M_E,descuento):
    # sumatorio_cons_mens=cons_mens_P1+cons_mens_P2+cons_mens_P3
    precio_P1_descuento= P1M_E * (1-descuento) #€
    precio_P2_descuento= P2M_E * (1-descuento)
    precio_P3_descuento= P3M_E * (1-descuento)
    
    #añadir P4,P5,P6 como None/0/1

    total_pago_P1_energia= cons_anual_P1 * precio_P1_descuento #€
    total_pago_P2_energia= cons_anual_P2 * precio_P2_descuento
    total_pago_P3_energia= cons_anual_P3 * precio_P3_descuento

    sumatorio_total_pago_energia = total_pago_P1_energia + total_pago_P2_energia + total_pago_P3_energia
    return sumatorio_total_pago_energia

def calcular_potencia(potencia_contratada_P1,potencia_contratada_P2,dias,precio_potencia_dia_P1,precio_potencia_dia_P2,dto_p):
    total_pago_P1_potencia= dias * precio_potencia_dia_P1 * potencia_contratada_P1 * (1-dto_p)
    total_pago_P2_potencia= dias * precio_potencia_dia_P2 * potencia_contratada_P2 * (1-dto_p)
    sumatorio_total_pago_potencia = total_pago_P1_potencia + total_pago_P2_potencia
    return sumatorio_total_pago_potencia
    
def calcular_total_factura(sumatorio_total_pago_energia,sumatorio_total_pago_potencia,impuesto_electrico,otros,alquiler_equipo,IVA):
    bi_IVA= (sumatorio_total_pago_energia + sumatorio_total_pago_potencia
         +impuesto_electrico*12 + otros*12 + alquiler_equipo*12)
    importe_total_factura_anual= bi_IVA * (1+IVA)
    return importe_total_factura_anual


pd.set_option('display.max_rows', None)

def encontrar_opcion_mas_barata_anual(df_energia, df_potencia):
    opciones = []

    df_combinado = pd.merge(df_energia, df_potencia, on='CIA',suffixes=["_E","_P"])
    df_combinado.dropna(axis=0,inplace=True)

    cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834, 955, 1824 #CAMBIAR A INPUT DE LA FACTURA
    potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #CAMBIAR A INPUT DE LA FACTURA
    precio_media_P1, precio_media_P2, precio_media_P3 = 0.128515, 0.128515, 0.128515#---->esto tendria q ser la media de los p de 12 fras
    precio_potencia_dia_P1, precio_potencia_dia_P2 = 0.09968, 0.033816
    descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
    dias = 365 # CAMBIAR A INPUT DE LA FACTURA
    dto_p = 0.0
    
    sumatorio_total_pago_energia = calcular_energia(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_media_P1, precio_media_P2, precio_media_P3, descuento)
    sumatorio_total_pago_potencia = calcular_potencia(potencia_contratada_P1, potencia_contratada_P2, dias, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
    impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

    # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

    importe_total_factura_anual_actual = round(calcular_total_factura(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)





    for index,row in df_combinado.iterrows():
        cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834, 955, 1824 #CAMBIAR A INPUT DE LA FACTURA
        potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #CAMBIAR A INPUT DE LA FACTURA
        precio_media_P1, precio_media_P2, precio_media_P3 = row['P1M_E'], row['P2M_E'], row['P3M_E']
        precio_potencia_dia_P1, precio_potencia_dia_P2 = row['P1'], row['P2']
        descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
        dias = 365 # CAMBIAR A INPUT DE LA FACTURA
        dto_p = 0.0

        sumatorio_total_pago_energia = calcular_energia(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_media_P1, precio_media_P2, precio_media_P3, descuento)
        sumatorio_total_pago_potencia = calcular_potencia(potencia_contratada_P1, potencia_contratada_P2, dias, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
        impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

        # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

        importe_total_factura_anual_ppta = round(calcular_total_factura(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)
        
        opciones.append({
            'CIA': row['CIA'],
            'FEE': row['FEE'],
            'PRODUCTO_CIA': row['PRODUCTO_CIA'],
            'CostoTotal': importe_total_factura_anual_ppta
        })

    # Opción más barata para cada compañía
    df_opciones = pd.DataFrame(opciones)
    idx_opcion_mas_barata = df_opciones['CostoTotal'].idxmin()
    
    opcion_barata=df_opciones.iloc[idx_opcion_mas_barata]

    opciones_mas_baratas = df_opciones.nsmallest(5, 'CostoTotal')
    ahorro_euros=round(importe_total_factura_anual_actual-opcion_barata['CostoTotal'],2)
    porcentaje_ahorro= round((ahorro_euros/importe_total_factura_anual_actual)*100,2)

    return ('Precio actual:',importe_total_factura_anual_actual,'Opción más barata:',opcion_barata,
            'Ahorro:',ahorro_euros, 'Porcentaje de ahorro:',f"{porcentaje_ahorro:.1f}%","Opciones más baratas:", opciones_mas_baratas)




encontrar_opcion_mas_barata_anual(df_medindx12_penins_2,index_power_filtrado)



def encontrar_opcion_mas_barata_cia_anual(df_energia, df_potencia):
    opciones = []

    
    df_combinado = pd.merge(df_energia, df_potencia, on='CIA',suffixes=["_E","_P"])
    df_combinado.dropna(axis=0,inplace=True)

    cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834, 955, 1824 #CAMBIAR A INPUT DE LA FACTURA
    potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #CAMBIAR A INPUT DE LA FACTURA
    precio_media_P1, precio_media_P2, precio_media_P3 = 0.128515, 0.128515, 0.128515#---->esto tendria q ser la media de los p de 12 fras
    precio_potencia_dia_P1, precio_potencia_dia_P2 = 0.09968, 0.033816
    descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
    dias = 365 # CAMBIAR A INPUT DE LA FACTURA
    dto_p = 0.0
    
    sumatorio_total_pago_energia = calcular_energia(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_media_P1, precio_media_P2, precio_media_P3, descuento)
    sumatorio_total_pago_potencia = calcular_potencia(potencia_contratada_P1, potencia_contratada_P2, dias, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
    impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

    # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

    importe_total_factura_anual_actual = round(calcular_total_factura(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)
    

# Create a dictionary to store the minimum cost for each 'CIA'
    min_cost_dict = {}

    for index,row in df_combinado.iterrows():
        cons_anual_P1, cons_anual_P2, cons_anual_P3 = 834, 955, 1824 #CAMBIAR A INPUT DE LA FACTURA
        potencia_contratada_P1, potencia_contratada_P2 = 3.45, 3.45 #CAMBIAR A INPUT DE LA FACTURA
        precio_media_P1, precio_media_P2, precio_media_P3 = row['P1M_E'], row['P2M_E'], row['P3M_E']
        precio_potencia_dia_P1, precio_potencia_dia_P2 = row['P1'], row['P2']
        descuento = 0.02  # CAMBIAR A INPUT DE LA FACTURA
        dias = 365 # CAMBIAR A INPUT DE LA FACTURA
        dto_p = 0.0

        sumatorio_total_pago_energia = calcular_energia(cons_anual_P1, cons_anual_P2, cons_anual_P3, precio_media_P1, precio_media_P2, precio_media_P3, descuento)
        sumatorio_total_pago_potencia = calcular_potencia(potencia_contratada_P1, potencia_contratada_P2, dias, precio_potencia_dia_P1, precio_potencia_dia_P2,dto_p)
        impuesto_electrico, otros, alquiler_equipo, IVA = 0.33, 1.27, 0.88,0.05  # CAMBIAR A INPUT DE LA FACTURA

        # print(sumatorio_total_pago_energia,sumatorio_total_pago_potencia)

        importe_total_factura_anual_ppta = round(calcular_total_factura(sumatorio_total_pago_energia, sumatorio_total_pago_potencia, impuesto_electrico, otros,alquiler_equipo, IVA),2)

  

    

    

        # Update the minimum cost for each 'CIA'
        if row['CIA'] not in min_cost_dict or importe_total_factura_anual_ppta < min_cost_dict[row['CIA']]:
            min_cost_dict[row['CIA']] = importe_total_factura_anual_ppta

    
    # Create a list with the minimum cost entries for each 'CIA' and calculate savings for each option
    opciones = [{
        'CIA': cia,
        'FEE': df_combinado.loc[df_combinado['CIA'] == cia, 'FEE'].values[0],
        'PRODUCTO_CIA': df_combinado.loc[df_combinado['CIA'] == cia, 'PRODUCTO_CIA'].values[0],
        'CostoTotal': min_cost,
        'Ahorro': round(importe_total_factura_anual_actual - min_cost, 2),
        'PorcentajeAhorro': round(((importe_total_factura_anual_actual - min_cost) / importe_total_factura_anual_actual) * 100, 2)
    } for cia, min_cost in min_cost_dict.items()]

    # Convert the list to a DataFrame
    df_opciones = pd.DataFrame(opciones)

    # Opción más barata para cada compañía
    opciones_mas_baratas = df_opciones.nsmallest(5, 'CostoTotal')

    return ('Precio actual:', importe_total_factura_anual_actual, 'Opciones más baratas:', opciones_mas_baratas)







encontrar_opcion_mas_barata_cia_anual(df_medindx12_penins_2,index_power_filtrado)

df_medindx12_penins_2
